% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{display_all_palettes}
\alias{display_all_palettes}
\title{Display all palettes}
\usage{
display_all_palettes(n = NULL)
}
\arguments{
\item{n}{Number of palettes to display. All palettes will be displayed in default}
}
\description{
Display all palettes
}
\examples{
display_all_palettes()
}
