% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{paint_palette}
\alias{paint_palette}
\title{Paintings Palette Generator}
\usage{
paint_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choice are:
\code{Pearlgirl}, \code{Splash}, \code{Autumn},
\code{Villeneuve}, \code{Ophelia}, \code{Kitchen},
\code{Spring}, \code{Strawberries}, \code{Seascape},
\code{Twilight}, \code{Abstract}, \code{Vesuvius}}

\item{n}{Number of colors you want.}

\item{type}{Use "discrete" or "continuous". Use "continuous" to automatically
interpolate between colors if you want more colors.}
}
\value{
A vector of colors.
}
\description{
Paintings Palette Generator
}
\examples{
paint_palette("Ophelia")
paint_palette("Ophelia", n=3)
paint_palette("Autumn", n=100, type="continuous")
}
\keyword{colors}
\keyword{palette}
