% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{painting_palettes}
\alias{painting_palettes}
\title{Complete List of Palettes}
\format{
An object of class \code{list} of length 12.
}
\usage{
painting_palettes
}
\description{
Use \code{\link{paint_palette}} to construct palettes of desired length.
}
\keyword{datasets}
